requirements.txt
mysql-connector-python 
pandas
stdiomask
colorama 
pymysql
sqlalchemy
tabulate
